create constructor module
